The superkey sort wasn't working so I had to go a different direction.  Here are some notes on my testing of this commit.
The combined count of the top ten words without filtering of the control list is 49940.
I ran the input document through a word processor and did a word count and it came up with 215824.
The unique word count I am getting in the program right now is 18922.  At first glance, I thought these numbers did not add up.
Upon further analysis, I found that that 18922 number only counts each word (it is a dictionary key) once.